+++ src/reproductive_rights_data_project/api/github/open_data_se.py
@@ -3,7 +3,7 @@ This file interfaces with the OpenDataSE GitHub repository.
 """
 
 import requests
-from reproductive_rights_data_project.util.constants import REQUEST_TIMEOUT
+from src.reproductive_rights_data_project.util.constants import REQUEST_TIMEOUT
 
 
 def get_state_zip_code_geo_json(state_abbrev, state_name):
