+++ src/reproductive_rights_data_project/data_handling/ansirh/clean.py
@@ -4,7 +4,7 @@ Clean ANSIRH data to be munged.
 
 import re
 from math import isnan
-from reproductive_rights_data_project.util.constants import (
+from src.reproductive_rights_data_project.util.constants import (
     FILTERED_CHARACTERS_REGEX,
     TYPE_DEFAULTS,
 )
