+++ src/reproductive_rights_data_project/test/test_api.py
@@ -6,10 +6,10 @@ import json
 from http import HTTPStatus
 import responses
 
-from reproductive_rights_data_project.api.github.open_data_se import (
+from src.reproductive_rights_data_project.api.github.open_data_se import (
     get_state_zip_code_geo_json,
 )
-from reproductive_rights_data_project.api.abortion_policy_api import (
+from src.reproductive_rights_data_project.api.abortion_policy_api import (
     get_data,
     add_missing_states,
     set_default_types,
