+++ src/reproductive_rights_data_project/test/test_data_handling.py
@@ -4,12 +4,13 @@ This file contains all testing functions for the data_handling package.
 
 import pandas as pd
 from math import isnan
-from reproductive_rights_data_project.util.constants import STANDARD_ENCODING
-from reproductive_rights_data_project.data_handling.ansirh.clean import (
+from src.reproductive_rights_data_project.util.constants import \
+    STANDARD_ENCODING
+from src.reproductive_rights_data_project.data_handling.ansirh.clean import (
     clean,
     set_default_types,
 )
-from reproductive_rights_data_project.data_handling.ansirh.process import (
+from src.reproductive_rights_data_project.data_handling.ansirh.process import (
     split_by_state,
     translate_code_to_state,
     split_by_zip,
