+++ src/reproductive_rights_data_project/test/test_util.py
@@ -1,8 +1,9 @@
 """
 This file contains all testing functions for the util package.
 """
-from reproductive_rights_data_project.util.constants import STANDARD_ENCODING
-from reproductive_rights_data_project.util.functions import to_json
+from src.reproductive_rights_data_project.util.constants import \
+    STANDARD_ENCODING
+from src.reproductive_rights_data_project.util.functions import to_json
 import json
 import os
