+++ src/reproductive_rights_data_project/visualization/functions.py
@@ -3,7 +3,7 @@ This file contains helper functions for generating data visualizations.
 """
 
 import json
-from reproductive_rights_data_project.util.constants import (
+from src.reproductive_rights_data_project.util.constants import (
     FILE_NAME_ANSIRH_CLEAN_DATA,
     STANDARD_ENCODING,
 )
@@ -54,7 +54,7 @@ def get_city_clinic_counts():
                     continue
                 city_state = ", ".join([clinic["city"].title(), state])
                 count_city_clinics[city_state] = (
-                    count_city_clinics.get(city_state, 0) + 1
+                        count_city_clinics.get(city_state, 0) + 1
                 )
 
     # sort by number of clinics
