+++ src/reproductive_rights_data_project/visualization/maps/usa_country.py
@@ -6,10 +6,11 @@ import plotly.express as px
 import pandas as pd
 import json
 
-from reproductive_rights_data_project.visualization.abstract_visualization import (
+from src.reproductive_rights_data_project.visualization.abstract_visualization import (
     Visualization,
 )
-from reproductive_rights_data_project.util.constants import STANDARD_ENCODING
+from src.reproductive_rights_data_project.util.constants import \
+    STANDARD_ENCODING
 
 
 class USAMap(Visualization):
@@ -21,10 +22,10 @@ class USAMap(Visualization):
     """
 
     def __init__(
-        self,
-        gestational_info_file_name,
-        locations_file_name,
-        state_abbrevs_file_name,
+            self,
+            gestational_info_file_name,
+            locations_file_name,
+            state_abbrevs_file_name,
     ):
         """
         Author(s): Michael Plunkett
@@ -45,7 +46,7 @@ class USAMap(Visualization):
         """
 
         with open(
-            self._gestational_info_file_name, encoding=STANDARD_ENCODING
+                self._gestational_info_file_name, encoding=STANDARD_ENCODING
         ) as gestational, open(
             self._locations_file_name, encoding=STANDARD_ENCODING
         ) as locations, open(
